backend/app/requirements.txt
# === Core ===
fastapi>=0.115
uvicorn[standard]>=0.30

# Database + async driver
SQLAlchemy>=2.0
asyncpg>=0.29

# Config / settings (.env)
pydantic>=2.7
pydantic-settings>=2.2
python-dotenv>=1.0

# Auth
passlib[bcrypt]>=1.7
PyJWT>=2.8

# Forms / uploads (корисно навіть якщо зараз не використовуєш)
python-multipart>=0.0.9

# Utils / perf (опційно)
orjson>=3.10        # швидкий JSON-енкодер; FastAPI сам підхопить, якщо встановлено
email-validator>=2.1 # валідація емейлів у схемах (за бажанням)
httpx>=0.27          # на майбутнє: вебхуки/зовнішні запити
# slowapi>=0.1.9     # rate limiting (опційно, коли додаватимемо ліміти)
